requirements.txt
appdirs==1.4.3
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
Deprecated==1.2.10
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
py==1.9.0
PyGithub==1.53
PyJWT==1.7.1
pyparsing==2.4.6
pyspark==3.0.0
pytest==6.1.2
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
